backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pytesseract==0.3.10
pdf2image==1.16.3
google-cloud-aiplatform>=1.38.0
google-generativeai
httpx>=0.23.0,<1.0.0
fpdf2==2.7.6
pillow>=10.1.0
pypdf2==3.0.1
python-dotenv==1.0.0
pandas>=2.0.0
openpyxl>=3.1.0
python-docx>=1.1.0
celery>=5.3.0
redis>=5.0.0

